requirements.txt
streamlit>=1.42.0
streamlit-mermaid
supabase>=2.11.0
boto3>=1.36.11
pandas>=2.2.3
Authlib>=1.3.2
openpyxl
xlrd
python-pptx
st-supabase-connection
cleanco
pydantic
pydantic-ai
nest_asyncio
tavily-python
google-genai
apify-client
python-dotenv
scikit-learn
numpy
google-generativeai
elevenlabs
streamlit-anywidget
gtts
pydub
linkup-sdk
trafilatura
graphiti-core[google-genai]
networkx
neo4j
matplotlib
pyvis
pymupdf
python-docx
qdrant_client
docx2pdf
fastembed
llama-index
llama-index-vector-stores-qdrant
llama-index-readers-file
llama-index-embeddings-fastembed
llama-index-llms-openai
wordcloud
qrcode
html2image